Build helpers unittests and functional
